Revert "Rental Properties has foreign key to personal info rather than rentalUser"
This reverts commit c7e48463fefb1eca2d92305f84865755958c46c3.

+++ home-rental-app/src/main/java/com/rental/service/RentalPropsService.java
@@ -29,7 +29,7 @@ public class RentalPropsService {
 		this.rpr.save(rentalProp);
 	}
 	
-	public Optional<RentalProps> getRentalPropertyById(Long id) {
+	public Optional<RentalProps> getRentalPropertyById(Integer id) {
 		return this.rpr.findById(id);
 	}
